Добавление новых Swift-файлов, связывание класса MovieQuizViewController с другими
