+++ MovieQuiz/Models/QuizQuestion.swift
@@ -1,0 +1,15 @@
+//
+//  QuizQuestion.swift
+//  MovieQuiz
+//
+//  Created by Дмитрий Герасимов on 21.08.2023.
+//
+
+import Foundation
+
+// Структура для конвертации вопросов с мока
+ struct QuizQuestion {
+    let image: String
+    let text: String
+    let correctAnswer: Bool
+}
