+++ MovieQuiz/Models/QuizResultsViewModel.swift
@@ -1,0 +1,15 @@
+//
+//  QuizResultsViewModel.swift
+//  MovieQuiz
+//
+//  Created by Дмитрий Герасимов on 21.08.2023.
+//
+
+import Foundation
+
+// Структура для работы алерта
+ struct QuizResultsViewModel{
+    let title: String
+    let text: String
+    let buttonText: String
+}
