+++ MovieQuiz/Models/QuizStepViewModel.swift
@@ -1,0 +1,16 @@
+//
+//  QuizStepViewModel.swift
+//  MovieQuiz
+//
+//  Created by Дмитрий Герасимов on 21.08.2023.
+//
+
+import Foundation
+import UIKit
+
+// Вью модель для состояния "вопрос показан"
+ struct QuizStepViewModel{
+    let image: UIImage
+    let question: String
+    let questionNumber: String
+}
