+++ MovieQuiz/Presentation/MovieQuizViewController.swift
@@ -15,68 +15,11 @@ final class MovieQuizViewController: UIViewController {
     // Счетчик правильных ответов
     private var correctAnswers: Int = 0
     
-    // Структура для конвертации вопросов с мока
-    private struct QuizQuestion {
-        let image: String
-        let text: String
-        let correctAnswer: Bool
-    }
-    // Вью модель для состояния "вопрос показан"
-    private struct QuizStepViewModel{
-        let image: UIImage
-        let question: String
-        let questionNumber: String
-    }
-    // Структура для работы алерта
-    private struct QuizResultsViewModel{
-        let title: String
-        let text: String
-        let buttonText: String
-    }
-    // Mock-данные для квиза
-    private let questions: [QuizQuestion] = [
-        QuizQuestion(
-            image:"The Godfather",
-            text: "Рейтинг этого фильма больше чем 6?",
-            correctAnswer: true),
-        QuizQuestion(
-            image: "The Dark Knight",
-            text: "Рейтинг этого фильма больше чем 6?",
-            correctAnswer: true),
-        QuizQuestion(
-            image: "Kill Bill",
-            text: "Рейтинг этого фильма больше чем 6?",
-            correctAnswer: true),
-        QuizQuestion(
-            image: "The Avengers",
-            text: "Рейтинг этого фильма больше чем 6?",
-            correctAnswer: true),
-        QuizQuestion(
-            image: "Deadpool",
-            text: "Рейтинг этого фильма больше чем 6?",
-            correctAnswer: true),
-        QuizQuestion(
-            image: "The Green Knight",
-            text: "Рейтинг этого фильма больше чем 6?",
-            correctAnswer: true),
-        QuizQuestion(
-            image: "Old",
-            text: "Рейтинг этого фильма больше чем 6?",
-            correctAnswer: false),
-        QuizQuestion(
-            image: "The Ice Age Adventures of Buck Wild",
-            text: "Рейтинг этого фильма больше чем 6?",
-            correctAnswer: false),
-        QuizQuestion(
-            image: "Tesla",
-            text: "Рейтинг этого фильма больше чем 6?",
-            correctAnswer: false),
-        QuizQuestion(
-            image: "Vivarium",
-            text: "Рейтинг этого фильма больше чем 6?",
-            correctAnswer: false),
-        
-    ]
+    private let questionsAmount: Int = 10
+    private let questionFactory: QuestionFactory = QuestionFactory()
+    private var currentQuestion: QuizQuestion?
+    
+    
     
     
     
@@ -85,18 +28,24 @@ final class MovieQuizViewController: UIViewController {
     
     // Lifecycle
     override func viewDidLoad() {
-        let currentQuestion = questions[currentQuestionIndex]
-        imageView.layer.cornerRadius = 20
         super.viewDidLoad()
-        show(quiz: convert(model: currentQuestion))
+        imageView.layer.cornerRadius = 20
+        
+        if let firstQuestion = questionFactory.requestNextQuestion() {
+            currentQuestion = firstQuestion
+            let viewModel = convert(model: firstQuestion)
+            show(quiz: viewModel)
+        }
     }
+    
     // Метод для конвертации вопросов из мока
     private func convert(model: QuizQuestion) -> QuizStepViewModel{
         return QuizStepViewModel(
             image: UIImage(named: model.image) ?? UIImage(),
             question: model.text,
-            questionNumber: "\(currentQuestionIndex + 1)/\(questions.count)")
+            questionNumber: "\(currentQuestionIndex + 1)/\(questionsAmount)")
     }
+    
     // Метод для отображения вопросов
     private func show(quiz step: QuizStepViewModel){
         counterLabel.text = step.questionNumber
@@ -104,8 +53,8 @@ final class MovieQuizViewController: UIViewController {
         imageView.image = step.image
         self.yesButton.isEnabled = true
         self.noButton.isEnabled = true
-        
     }
+    
     // Метод для отображения рамки правильного или неправильного ответа
     private func showAnswerResult(isCorrect: Bool) {
         if isCorrect{
@@ -139,7 +88,7 @@ final class MovieQuizViewController: UIViewController {
     }
     // Метод логики перехода в один из сценариев
     private func showNextQuestionOrResults() {
-        if currentQuestionIndex == questions.count - 1{
+        if currentQuestionIndex == questionsAmount - 1{
             let text = "Ваш результат: \(correctAnswers) из 10"
             let viewModel = QuizResultsViewModel(
                 title: "Этот раунд окончен!",
@@ -149,44 +98,21 @@ final class MovieQuizViewController: UIViewController {
             show(quiz: viewModel)
         }else{
             currentQuestionIndex += 1
-            let nexQuestion = questions[currentQuestionIndex]
-            let viewModel = convert(model: nexQuestion)
-            imageView.layer.borderWidth = 0
-            show(quiz: viewModel)
+            if let nextQuestion = questionFactory.requestNextQuestion() {
+                currentQuestion = nextQuestion
+                let viewModel = convert(model: nextQuestion)
+                
+                show(quiz: viewModel)
+            }
         }
-    }
-    // Метод работы с результатами квиза
-    private func show(quiz result: QuizResultsViewModel) {
-        let alert = UIAlertController(
-            title: result.title,
-            message: result.text,
-            preferredStyle: .alert)
-        
-        let action = UIAlertAction(title: result.buttonText, style: .default) { [weak self ] _ in
-            guard let self = self else { return }
-            self.currentQuestionIndex = 0
-            
-            // скидываем счётчик правильных ответов
-            self.correctAnswers = 0
-            
-            // заново показываем первый вопрос
-            let firstQuestion = self.questions[self.currentQuestionIndex]
-            let viewModel = self.convert(model: firstQuestion)
-            self.show(quiz: viewModel)
-        }
-        
-        alert.addAction(action)
-        
-        self.present(alert, animated: true, completion: nil)
-        
-        
     }
     
     // MARK: - Actions
-    
     // Кнопка ответа - "Нет"
-    @IBAction private func noButtonClicked(_ sender: UIButton) {
-        let currentQuestion = questions[currentQuestionIndex]
+    @IBAction private func noButtonClicked(_ sender: UIButton){
+        guard let currentQuestion = currentQuestion else {
+            return
+        }
         let givenAnswer = false
         showAnswerResult(isCorrect: givenAnswer == currentQuestion.correctAnswer)
     }
@@ -194,14 +120,48 @@ final class MovieQuizViewController: UIViewController {
     
     // Кнопка ответа - "Да"
     @IBAction private func yesButtonClicked(_ sender: UIButton) {
-        let currentQuestion = questions[currentQuestionIndex]
+        guard let currentQuestion = currentQuestion else {
+            return
+        }
         let givenAnswer = true
         showAnswerResult(isCorrect: givenAnswer == currentQuestion.correctAnswer)
     }
+   
+   
+        // Метод работы с результатами квиза
+    private func show(quiz result: QuizResultsViewModel) {
+        let alert = UIAlertController(
+            title: result.title,
+            message: result.text,
+            preferredStyle: .alert)
+        
+        let action = UIAlertAction(title: result.buttonText, style: .default)  { [weak self ] _ in
+            guard let self = self else { return }
+            self.currentQuestionIndex = 0
+            
+            // скидываем счётчик правильных ответов
+            self.correctAnswers = 0
+            
+            // заново показываем первый вопрос
+            if let firstQuestion = self.questionFactory.requestNextQuestion() {
+                self.currentQuestion = firstQuestion
+                let viewModel = self.convert(model: firstQuestion)
+                
+                self.show(quiz: viewModel)
+            }
+        }
+            alert.addAction(action)
+            
+            self.present(alert, animated: true, completion: nil)
+        }
+    }
+            
+            
+           
+            
+        
     
     
-}
-
 
 /*
  Mock-данные
